requirements.txt
psycopg2-binary==2.9.11
requests==2.32.5
pyyaml==6.0.1
rich==13.7.1
